requirements.txt
# Core dependencies
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Trading and analysis
ccxt>=4.0.0
# Using alternative TA libraries instead of TA-Lib
ta>=0.10.0
pandas-ta>=0.3.0
yfinance>=0.2.0

# Data sources
tweepy>=4.14.0
praw>=7.7.0
python-telegram-bot>=20.0

# Web and APIs
flask>=2.3.0
flask-cors>=4.0.0
websocket-client>=1.6.0

# Database (sqlite3 is built into Python, no need to install)
sqlalchemy>=2.0.0

# Utilities
schedule>=1.2.0
colorlog>=6.7.0
python-dateutil>=2.8.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Phase 3 ML/Analytics Requirements
scikit-learn>=1.3.0
joblib>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.17.0

# Optional Enhanced ML Features (uncomment if needed)
# Advanced NLP/Sentiment Analysis
# transformers>=4.21.0
# torch>=2.0.0
# textblob>=0.17.0

# Enhanced Technical Analysis (uncomment if needed)
# talib>=0.4.0  # Requires manual C++ compilation on some systems

# Machine Learning Extras (uncomment for advanced features)
# xgboost>=1.7.0
# lightgbm>=4.0.0
# scipy>=1.10.0pycoingecko==3.1.0
pytrends==4.9.2
